+++ front/src/context/AuthContext.jsx
@@ -16,7 +16,6 @@ export function AuthProvider({ children }) {
 
   useEffect(() => {
     const unsubscribed = auth.onAuthStateChanged((user) => {
-      console.log(user)
       setUser(user);
     });
     return () => {
